+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - Added option to set `--cache-freq`.
 - Added option to set `--zap-log-level`.
+- Added more data fields to existing collections -- expanded node and job
+  states, memory usage.
 
 ### Fixed